test2

+++ get-tl.ts
@@ -97,6 +97,12 @@ export default async function getTL(arg: string, opt: Array<string>){
             }
 
             let response = await inst.get(`/lol/match/v5/matches/${region}_${ID}/timeline`);
+
+            if (fs.existsSync(`${files[i].parentPath.replace('overview', 'timeline')}/${files[i].name.replace('overview', 'timeline')}`)){
+                console.log(`timeline_${region}_${ID} alredy exists!`);
+                continue;
+            }
+            
             fs.writeFileSync(`${files[i].parentPath.replace('overview', 'timeline')}/${files[i].name.replace('overview', 'timeline')}`,'')
 
             console.log(`Game ${response.data['metadata']['matchId']} Found! Found!(${i}/${files.length}) [${Math.floor(i/(files.length/100))}%]`);
